--- src/main/java/domain/Money.java
@@ -1,8 +1,0 @@
-package domain;
-
-public record Money(long amount) {
-
-    public Money multiply(double factor) {
-        return new Money(Math.round(amount * factor));
-    }
-}
