+++ src/main/java/domain/betting/Betting.java
@@ -1,6 +1,6 @@
 package domain.betting;
 
-import domain.Money;
+import domain.money.Money;
 import domain.participant.Player;
 
 public record Betting(Player player, Money money) {
